+++ mod.ts
@@ -37,6 +37,7 @@ export { EdgeElement } from "./edge/bob.tsx";
 export { build } from "./core/api.ts";
 export { serve } from "./core/server.ts";
 export { h } from "./core/jsx.ts";
+export { bob } from "./core/cli.ts";
 export * from "./utils/get-path.ts";
 export * from "./utils/read-contents.ts";
 export * from "./utils/sort-date.ts";